Add relative positioning to t2-chest command
Uses '~' to indicate a relative position compared to the issuer of the command (usually the player).  Example: t2-chest ~ ~1 ~-1 would put it at the same x as the player, y+1 as the player and z-1 as the player.  This does not interfere with the use of absolute positioning as originally written.